Extract LIMIT tests to separate file
